feat: move output processing to handler
